2010-05-27	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libnetworking/netinet/raw_ip.c,
	libnetworking/netinet/udp_usrreq.c:
	Cast mbufs to uintptr_t instead of u_long.

+++ cpukit/libnetworking/netinet/raw_ip.c
@@ -336,7 +336,7 @@ rip_usrreq(struct socket *so, int req, struct mbuf *m, struct mbuf *nam,
 	int s;
 
 	if (req == PRU_CONTROL)
-		return (in_control(so, (u_long)m, (caddr_t)nam,
+		return (in_control(so, (uintptr_t)m, (caddr_t)nam,
 			(struct ifnet *)control));
 
 	switch (req) {
